requirements.txt
Django==3.2
djangorestframework
psycopg2-binary==2.8.6
Pillow

